+++ src/cli/_validators.py
@@ -8,6 +8,7 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple, Union
 
 import click
+import docker
 import yaml
 
 from ._common import existing_cluster_ids
@@ -167,36 +168,55 @@ def validate_volumes(
     ctx: click.core.Context,
     param: Union[click.core.Option, click.core.Parameter],
     value: Any,
-) -> Dict[str, Dict[str, str]]:
+) -> List[docker.types.Mount]:
     """
-    Turn volume definition strings into dictionaries that ``docker-py`` can
-    use.
+    Turn volume definition strings into ``Mount``s that ``docker-py`` can use.
     """
     for _ in (ctx, param):
         pass
-    volumes = {}
+    mounts = []
     for volume_definition in value:
         parts = volume_definition.split(':')
 
         if len(parts) == 1:
             host_src = str(uuid.uuid4())
             [container_dst] = parts
-            mode = 'rw'
+            read_only = False
         elif len(parts) == 2:
             host_src, container_dst = parts
-            mode = 'rw'
+            read_only = False
         elif len(parts) == 3:
             host_src, container_dst, mode = parts
+            if mode == 'ro':
+                read_only = True
+            elif mode == 'rw':
+                read_only = False
+            else:
+                message = (
+                    'Mode in "{volume_definition}" is "{mode}". '
+                    'If given, the mode must be one of "ro", "rw".'
+                ).format(
+                    volume_definition=volume_definition,
+                    mode=mode,
+                )
+                raise click.BadParameter(message=message)
         else:
             message = (
+                '"{volume_definition}" is not a valid volume definition. '
                 'See '
                 'https://docs.docker.com/engine/reference/run/#volume-shared-filesystems '  # noqa: E501
                 'for the syntax to use.'
-            )
+            ).format(volume_definition=volume_definition)
             raise click.BadParameter(message=message)
 
-        volumes[host_src] = {'bind': container_dst, 'mode': mode}
-    return volumes
+        mount = docker.types.Mount(
+            source=host_src,
+            target=container_dst,
+            type='bind',
+            read_only=read_only,
+        )
+        mounts.append(mount)
+    return mounts
 
 
 def validate_ovpn_file_does_not_exist(